% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{save_as_image}
\alias{save_as_image}
\title{save to image code that generates a plot}
\usage{
save_as_image(x, device, file = tempfile())
}
\arguments{
\item{x}{code to evaluate}

\item{device}{device to use to record image}

\item{file}{filepath for device}
}
\description{
save to image code that generates a plot
}
